+++ src/TestUtilities/EventStager.php
@@ -2,7 +2,11 @@
 
 declare(strict_types=1);
 
-namespace EventSauce\EventSourcing;
+namespace EventSauce\EventSourcing\TestUtilities;
+
+use EventSauce\EventSourcing\AggregateRootId;
+use EventSauce\EventSourcing\AggregateRootRepository;
+use EventSauce\EventSourcing\InMemoryMessageRepository;
 
 class EventStager
 {